Breaking changes
This commit includes a lot of breaking changes, so the major version number will be incremented.

The CerberusObjects must be manually registered on the framework using the checkIn() call (and checkOut() to unregister)
leaving to the application the choice of not share an object.

The HTTPClient and HTTPData classes have been rewritten to be more clear and usable.

All the main slot types have been implemented, more will be available in the future.

A Thread Continuos timing mode has been added to make threads run continuosly.

The Socket is now resettable (the system fd is renewed but the Socket instance remains the same, this can be handy sometimes,
e.g. when a TCP connection is closed but the application wants to reconnect without deleting and recreating the Socket instance)

General bug fixes, improvements and optimization changes